refactor: DTO 리팩토링
