+++ src/main/java/org/indoles/memberserviceserver/core/dto/SignInInfo.java
@@ -1,12 +1,13 @@
-package org.indoles.memberserviceserver.core.dto.response;
+package org.indoles.memberserviceserver.core.dto;
 
-import org.indoles.memberserviceserver.core.entity.enums.Role;
-import org.indoles.memberserviceserver.core.entity.exception.MemberException;
-import org.indoles.memberserviceserver.core.entity.exception.MemberExceptionCode;
+import org.indoles.memberserviceserver.core.domain.enums.Role;
+import org.indoles.memberserviceserver.global.exception.BadRequestException;
+import org.indoles.memberserviceserver.global.exception.ErrorCode;
 
 import java.util.Objects;
 
 public record SignInInfo(Long id, Role role) {
+    private static final String ERROR_NULL_VALUE = "%s는 Null일 수 없습니다.";
 
     public SignInInfo {
         validateNotNull(id, "로그인한 사용자의 식별자");
@@ -15,12 +16,12 @@ public record SignInInfo(Long id, Role role) {
 
     private void validateNotNull(Object value, String fieldName) {
         if (Objects.isNull(value)) {
-            throw new MemberException(MemberExceptionCode.FIELD_CANNOT_BE_NULL, fieldName);
+            throw new BadRequestException(String.format(ERROR_NULL_VALUE, fieldName), ErrorCode.G000);
         }
     }
 
     public boolean isType(Role role) {
         return this.role.equals(role);
     }
-
 }
+
