+++ src/main/java/org/indoles/memberserviceserver/core/dto/SignInResponseInfo.java
@@ -1,0 +1,9 @@
+package org.indoles.memberserviceserver.core.dto;
+
+import org.indoles.memberserviceserver.core.domain.enums.Role;
+
+public record SignInResponseInfo(
+        Role role
+) {
+}
+
