--- src/main/java/org/indoles/memberserviceserver/core/dto/request/MemberChargePointRequest.java
@@ -1,10 +1,0 @@
-package org.indoles.memberserviceserver.core.dto.request;
-
-import org.indoles.memberserviceserver.core.entity.enums.Role;
-
-public record MemberChargePointRequest(
-        Long memberId,
-        Role role,
-        Long amount
-) {
-}
